requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
colorama>=0.4.6
argparse>=1.4.0 